+++ packages/remix-dev/CHANGELOG.md
@@ -1,5 +1,13 @@
 # `@remix-run/dev`
 
+## 1.8.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.8.2`
+  - `@remix-run/serve@1.8.2`
+
 ## 1.8.1
 
 ### Patch Changes