+++ packages/remix-express/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@remix-run/express`
 
+## 1.8.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.8.2`
+
 ## 1.8.1
 
 ### Patch Changes